+++ checks/AZC/AZC001_client_missing_resource_manager_endpoint/AZC001.go
@@ -17,7 +17,7 @@ import (
 var Analyzer = &analysis.Analyzer{
 	Name:     "AZC001",
 	Doc:      "check for Azure SDK clients created without the resource manager endpoint explicitly specified via NewFoosClientWithBaseURI",
-	URL:      "https://github.com/katbyte/azproviderlint/blob/main/checks/AZC/AZC001_client_missing_base_uri/README.md",
+	URL:      "https://github.com/katbyte/azproviderlint/blob/main/checks/AZC/AZC001_client_missing_resource_manager_endpoint/README.md",
 	Requires: []*analysis.Analyzer{inspect.Analyzer},
 	Run:      run,
 }