+++ checks/AZS/AZS.go
@@ -7,9 +7,10 @@ import (
 	AZS001 "github.com/katbyte/azproviderlint/checks/AZS/AZS001_typed_sdk_model_64bit_types"
 	AZS002 "github.com/katbyte/azproviderlint/checks/AZS/AZS002_schema_default_type_mismatch"
 	AZS003 "github.com/katbyte/azproviderlint/checks/AZS/AZS003_schema_allows_empty_block"
-	AZS004 "github.com/katbyte/azproviderlint/checks/AZS/AZS004_enum_validation_missing_values"
+	AZS004 "github.com/katbyte/azproviderlint/checks/AZS/AZS004_enum_validation_possible_values"
 	AZS005 "github.com/katbyte/azproviderlint/checks/AZS/AZS005_resource_missing_data_source"
 	AZS006 "github.com/katbyte/azproviderlint/checks/AZS/AZS006_data_source_missing_properties"
+	AZS007 "github.com/katbyte/azproviderlint/checks/AZS/AZS007_optional_computed_missing_comment"
 )
 
 // Checks contains all AZS (schema & typed SDK model) analyzers.
@@ -20,4 +21,5 @@ var Checks = []*analysis.Analyzer{
 	AZS004.Analyzer,
 	AZS005.Analyzer,
 	AZS006.Analyzer,
+	AZS007.Analyzer,
 }