+++ checks/AZS/AZS007_optional_computed_missing_comment/AZS007_test.go
@@ -1,0 +1,21 @@
+package AZS007
+
+import (
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func TestAZS007(t *testing.T) {
+	t.Parallel()
+
+	_, filename, _, _ := runtime.Caller(0)
+	dir := filepath.Join(filepath.Dir(filename), "testdata")
+
+	analysistest.Run(t, dir, Analyzer, "azs007")
+
+	excludePackages = "migration"
+	analysistest.Run(t, dir, Analyzer, "migration")
+}